+++ tests/analysis/test_phase_diagram.py
@@ -30,7 +30,7 @@ from pymatgen.analysis.phase_diagram import (
 )
 from pymatgen.core import Composition, DummySpecies, Element
 from pymatgen.core.entries import ComputedEntry
-from pymatgen.entries.entry_tools import EntrySet
+from pymatgen.analysis.compatibility.entry_tools import EntrySet
 from pymatgen.util.testing import TEST_FILES_DIR, MatSciTest
 
 TEST_DIR = f"{TEST_FILES_DIR}/analysis"